\name{CompositionalHDDA-package}
\alias{CompositionalHDDA-package}
\docType{package}
\title{
High Dimensional Discriminant Analysis with Compositional Data
}

\description{
High dimensional discriminant analysis (HDDA) for compositional data using the alpha-transformation is performed.
}

\details{
\tabular{ll}{
Package: \tab CompositionalHDDA\cr
Type: \tab Package\cr
Version: \tab 1.0 \cr
Date: \tab 2025-07-02\cr
License: \tab GPL-2\cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>
}

\references{
Bouveyron C. Girard S. and Schmid C. (2007). High Dimensional Discriminant Analysis. Communications in Statistics: Theory and Methods, 36(14): 2607--2623.

Bouveyron C. Celeux G. and Girard S. (2010). Intrinsic dimension estimation by maximum likelihood in probabilistic PCA. Technical Report 440372, Universite Paris 1 Pantheon-Sorbonne.

Berge L. Bouveyron C. and Girard S. (2012). HDclassif: An R Package for Model-Based Clustering and Discriminant Analysis of High-Dimensional Data. Journal of Statistical Software, 46(6).

Tsagris M.T., Preston S. and Wood A.T.A. (2011). A data-based power transformation
for compositional data. In Proceedings of the 4th Compositional Data Analysis Workshop,
Girona, Spain. https://arxiv.org/pdf/1106.1451.pdf
}

%\note{
%}


